+++ MegaManLofi/PlayerConfig.h
@@ -8,10 +8,10 @@ namespace MegaManLofi
    class PlayerConfig
    {
    public:
-      Rectangle DefaultHitBox = { 0., 0., 0., 0. };
+      Rectangle DefaultHitBox = { 0, 0, 0, 0 };
 
-      double DefaultVelocityX = 0.;
-      double DefaultVelocityY = 0.;
+      long long DefaultVelocityX = 0;
+      long long DefaultVelocityY = 0;
 
       Direction DefaultDirection = (Direction)0;
    };
